requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
djangorestframework-simplejwt
django-filter
requests
celery
redis
drf-yasg
django-redis
whitenoise
black
